+++ src/main/scala/com/github/shadowsocks/services/LatencyTestService.scala
@@ -1,10 +1,10 @@
-package com.github.shadowsocks.service
+package com.github.shadowsocks.services
 
 import android.app.Service
 import android.content.Intent
 import android.os.IBinder
 
-class SubscriptionService extends Service {
+class LatencyTestService extends Service {
   override def onBind(intent: Intent): IBinder = null
 
   override def onStartCommand(intent: Intent, flags: Int, startId: Int): Int = {
